+++ CHANGELOG.md
@@ -9,7 +9,10 @@
 
 ## Fixed
 
+
 ## Changed
+* Fixed bug where incorrect gain was applied to the right CCD.
+* Changed the test for the gain, to correctly test that the left/right CCD have different gains.
 
 ## Added